Main View Fix - Refactor pearch product - Remove duplicated cards - Remove console logs - Update ProductList Class - Update utils renderListWithTemplate
